Got a trivial OSGi bundle running under Equinox!
You can see this working by running:

    ./scripts/run

It does not exit on its own, so you have to use ctrl-C to kill it.

This took hours of banging my head against the wall. This article
was by far the most helpful:

http://devblog.virtage.com/2012/08/osgi-with-equinox-for-beginners-overview/

I think the person who wrote this StackOverflow answer ripped off the previous article.
Unfortunately, it didn't link to that article, and did not go into as much detail,
but I found this one first:

http://stackoverflow.com/questions/17478544/java-jar-org-eclipse-osgi-3-9-0-v20130529-1710-jar-consolelog-unable-to-ac

This was somewhat helpful as an overview of Equinox, but the instructions did not work
out of the box, which was really frustrating, as they seemed to be very official:

https://www.eclipse.org/equinox/documents/quickstart-framework.php

This was helpful for explaining start levels:

http://eclipsesource.com/blogs/2009/06/10/osgi-and-start-levels/

This is ultimately more along the lines of what I am trying to do:

http://stackoverflow.com/questions/16692092/ijavaproject-without-eclipse-environment-in-jdt

We'll see if this is helpful now that I've gotten over the OSGi hurdle:

https://sdqweb.ipd.kit.edu/wiki/JDT_Tutorial:_Creating_Eclipse_Java_Projects_Programmatically